Fix missing work after rebase master
