--- packages/datastore/datastore/src/predicates/index.ts
@@ -1,3 +1,0 @@
-export * from "./Operators";
-export * from "./PredicateFunctions";
-export * from "./Predicates";
