+++ packages/datastore/datastore/src/storage/adapters/indexedDB/Predicate.ts
